consultations/2023-01-10/code/consultation_2023_01_10/src/Main.java
/* Дан массив из строк.
   Необходимо сформировать лист из строк, которые начинаются на заданную подстроку.

    [ "Hello world", "Hello Andy", "Java is the best", "Hell", "Say hello!", "hello Java", "HelloJava" ], "Hello" ->
    ("Hello world", "Hello Andy", "HelloJava")
    [ "Hello world", "Hello Andy", "Java is the best", "Hell", "Say hello!", "hello Java" ], "hello" ->
    ()
    [ "Hello world", "Hello Andy", "Java is the best", "Hell", "Say hello!", "hello Java" ], "  Hello    " ->
    ->()
    [],"Hello" -> ()
    null,"Hello" -> ()

 */

import java.util.ArrayList;

public class Main {
    public static void main(String[] args) {
        String[] strings = {};
        // ------

        getListOfStringStartWith(null,"hello");

    }

    public static ArrayList<String> getListOfStringStartWith(String[] strings, String prefix){

        ArrayList<String> res = new ArrayList<>();
        if(strings!=null) {
            for (int i = 0; i < strings.length; i++) {
                if (strings[i].startsWith(prefix)) {
                    res.add(strings[i]);
                }
            }
        }
        return res;

    }


}